+++ packages/volto-slate/src/utils/blocks.js
@@ -317,3 +317,15 @@ export const getAllBlocks = (properties, blocks) => {
   }
   return blocks;
 };
+
+export const clearFormatting = (editor) => {
+  const { slate } = config.settings;
+  Transforms.setNodes(editor, {
+    type: slate.defaultBlockType,
+  });
+  Transforms.unwrapNodes(editor, {
+    match: (n) => n.type && n.type !== slate.defaultBlockType,
+    mode: 'all',
+    split: false,
+  });
+};